agent/llm_client.py
"""
LLM Client

Handles communication with AI language models (OpenAI GPT-4, etc.).

Future implementation will include:
- OpenAI API integration
- HuggingFace API support
- Request/response handling
- Error handling and retries
- Token usage tracking
"""

import os
from dotenv import load_dotenv

load_dotenv()

class LLMClient:
    """Client for interacting with language models"""
    
    def __init__(self):
        self.api_key = os.getenv("OPENAI_API_KEY", "")
        self.model = os.getenv("AI_MODEL", "gpt-4")
        
    async def generate_response(self, prompt: str) -> str:
        """
        Generate a response from the LLM.
        
        Args:
            prompt: Input prompt for the LLM
            
        Returns:
            str: Generated response
        """
        # Placeholder for future implementation
        return "AI response placeholder"
